+++ bingmaps/__init__.py
@@ -1,0 +1,11 @@
+from .config import BING_MAPS_KEY
+
+from .__pkg__ import (
+    __author__,
+    __description__,
+    __email__,
+    __license__,
+    __package_name__,
+    __url__,
+    __version__
+)