+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Release 0.0.0
-*************
+=============
 
-Features:
-=========
+ - Location API services
+ - Finds location by address
+ - Finds location by point